third_party/blink/web_tests/editing/pasteboard/paste-noplugin-expected.txt

This test copies embed and object elements, pastes them into an editable area, and verifies that unsafe tags are not pasted when the setting prohibits this.

Original contents:
| "
"
| <p>
|   "Some text to paste."
| "
"
| <embed>
|   src="http://localhost:8000/evil.swf"
|   <shadow:root>
|     <slot>
| "
"
| <object>
|   data="http://localhost:8000/evil.swf"
|   <param>
|     name="src"
|     value="http://localhost:8000/evil.swf"
|   <embed>
|     src="http://localhost:8000/evil.swf"
|     <shadow:root>
|       <slot>
|   <shadow:root>
|     <slot>
| "
"

Pasted contents:
| <p>
|   "Some text to paste."
| " <#selection-caret>"
